requirements.txt
numpy>=1.18.1
portolan
pyyaml>=4.2b1
svgwrite
urllib3>=1.25.6
